:sparkles: feature: customizable colors
